Resistances primitive, item data (armor/helmets/ammo), serde tag
Add Resistances struct (all u32): ballistic, radiation, chemical,
thermal, electric, gravitational. Used by ArmorData, replaces
individual float fields. Ammo penetration changed to u32.

Add item JSON data:
- 6 suits: leather jacket → heavy rig
- 5 helmets: balaclava → CBRN mask
- 8 ammo types: 4 calibers × standard + AP

ItemData enum uses serde(tag = "type", rename_all = "lowercase")
for cleaner JSON: {"type": "ammo", "caliber": ...} instead of
{"Ammo": {"caliber": ...}}. Rarity gets rename_all lowercase.

+++ crates/cordon-core/src/item/data/ammo.rs
@@ -16,8 +16,11 @@ pub struct AmmoData {
     pub caliber: Id<Caliber>,
     /// Base damage per round.
     pub damage: f32,
-    /// Armor penetration value (0.0–1.0). Higher = better against armor.
-    pub penetration: f32,
+    /// Armor penetration value. Compared against armor's protection
+    /// value — if penetration exceeds armor, full damage goes through.
+    /// If lower, damage is reduced proportionally.
+    /// Typical range: 5 (pistol) to 40 (AP rifle).
+    pub penetration: u32,
     /// Number of rounds per box.
     pub quantity: u32,
 }
